fixed migration file

+++ db/migrate/20160308005231_create_schools.rb
@@ -3,8 +3,8 @@ class CreateSchools < ActiveRecord::Migration
     create_table :schools do |t|
       t.string :name
       t.string :address
-      t.double :lat
-      t.double :long
+      t.decimal :lat
+      t.decimal :long
 
       t.timestamps null: false
     end
